Fix for signed byte histogram failures

+++ cub/device/dispatch/device_histogram_dispatch.cuh
@@ -269,7 +269,7 @@ struct DeviceHistogramDispatch
             if (valid)
             {
                 bin = UpperBound(wrapped_levels, num_output_levels, (LevelT) sample) - 1;
-                valid = (bin < num_bins);
+                valid = (((unsigned int) bin) < num_bins);
             }
         }
     };
